# internal/controller/user.controller.go
package controller

import (
	"log"
	"rcontrisha/backend-eventhub/internal/dto"
	"rcontrisha/backend-eventhub/internal/service"
	"rcontrisha/backend-eventhub/pkg"

	"github.com/gin-gonic/gin"
	"github.com/gin-gonic/gin/binding"
)

type UserController struct {
	service *service.UserService
}

func NewUserController(service *service.UserService) *UserController {
	return &UserController{
		service: service,
	}
}

// User Profile
//
// @Summary				Get User Profile
// @Description		Retrieve the profile information of the authenticated user
// @Tags					user
// @Produce				json
// @Router				/user/profile	[get]
// @Security			BearerToken
// @Success				200		{object}	dto.Response
// @Failure				401		{object}	dto.Response
// @Failure				500		{object}	dto.Response
func (u *UserController) GetUserProfile(ctx *gin.Context) {
	token, exist := ctx.Get("token")
	if !exist {
		ctx.JSON(401, dto.Response{
			Status:  "failed",
			Message: "Token Data Not Found in Context.",
		})
	}

	claims := token.(pkg.JWTClaims)
	userId := claims.Id
	data, err := u.service.GetUserProfile(ctx, userId)
	if err != nil {
		ctx.JSON(500, dto.Response{
			Status:  "failed",
			Message: err.Error(),
			Data:    gin.H{},
		})
		return
	}

	ctx.JSON(200, dto.Response{
		Status:  "success",
		Message: "Successfully retrieve user's info.",
		Data: gin.H{
			"user": data,
		},
	})
}

// Change User Profile
//
// @Summary				Change User Profile
// @Description		Update the profile information of the authenticated user
// @Tags					user
// @Accept				json
// @Produce				json
// @Param					payload	body		dto.UserProfile	true	"User Profile Data"
// @Router				/user/profile	[patch]
// @Security			BearerToken
// @Success				200		{object}	dto.Response
// @Failure				401		{object}	dto.Response
// @Failure				500		{object}	dto.Response
func (u *UserController) ChangeUserProfile(ctx *gin.Context) {
	var payload dto.UserProfile
	if e := ctx.ShouldBindWith(&payload, binding.JSON); e != nil {
		log.Println("error", e.Error())
		ctx.JSON(500, gin.H{
			"success": false,
			"msg":     e.Error(),
		})
		return
	}

	token, exist := ctx.Get("token")
	if !exist {
		ctx.JSON(401, dto.Response{
			Status:  "failed",
			Message: "Token Data Not Found in Context.",
		})
	}

	claims := token.(pkg.JWTClaims)
	userId := claims.Id
	data, err := u.service.ChangeUserProfile(ctx, userId, payload)
	if err != nil {
		ctx.JSON(500, dto.Response{
			Status:  "failed",
			Message: err.Error(),
			Data:    gin.H{},
		})
		return
	}

	ctx.JSON(200, dto.Response{
		Status:  "success",
		Message: "Successfully update user's info.",
		Data: gin.H{
			"user": data,
		},
	})
}

// Change User Password
//
// @Summary				Change User Password
// @Description		Update the password of the authenticated user
// @Tags					user
// @Accept				json
// @Produce				json
// @Param					payload	body		dto.ChangePassword	true	"Change Password Data"
// @Router				/user/password	[patch]
// @Security			BearerToken
// @Success				200		{object}	dto.Response
// @Failure				401		{object}	dto.Response
// @Failure				500		{object}	dto.Response
func (u *UserController) ChangePassword(ctx *gin.Context) {
	var payload dto.ChangePassword
	if e := ctx.ShouldBindWith(&payload, binding.JSON); e != nil {
		log.Println("error", e.Error())
		ctx.JSON(500, gin.H{
			"success": false,
			"msg":     e.Error(),
		})
		return
	}

	token, exist := ctx.Get("token")
	if !exist {
		ctx.JSON(401, dto.Response{
			Status:  "failed",
			Message: "Token Data Not Found in Context.",
		})
	}

	claims := token.(pkg.JWTClaims)
	userId := claims.Id
	err := u.service.ChangeUserPwd(ctx, userId, payload)
	if err != nil {
		ctx.JSON(500, dto.Response{
			Status:  "failed",
			Message: err.Error(),
			Data:    gin.H{},
		})
		return
	}

	ctx.JSON(200, dto.Response{
		Status:  "success",
		Message: "Successfully update user's password.",
	})
}